Add new objects
After matches have been removed, any potential remaining objects
are moved down to replace the removed matches. After this point
we need to add the new objects to the tile map, using the following
steps:

Get a list of vacancies
Are any of these vacancies blocked?
If so, remove them from the vacancies list
With the remaining vacancies, identify their paths. Are there
any objects in the pathway?
If so, move them to the end of the path
Move in new objects
Update object information