Move objects if they are blocking empty tiles
When matches are removed, any remaining matches are moved
down vertically. This is fine in normal circumstances but
when obstacles are involved it can cause an issue.

This commit attempts to solve any problems by moving the
remaining objects if they are blocking empty cells that
new objects should be moved to.