Part 2: Move objects if they are blocking empty tiles
This is an addition to a previous commit. There was an issue because
I was correctly moving the blocked objectto the end of the path but
I was not taking into account any of the elements between the blocked
object and the path end.

Now the game identifies and concatenates each element's position
between the blocked object and the end of the path