Fix random bugs
Bug 1: Match element was referring to an off_screen object.
This was because the matches search for a matching location,
if there are two elements with the same location, it picks
the first one. It could be the wrong element.

Solution: When trying to identify matches only look at elements
that are on_screen.

Bug 2: User selected element referring to an off_screen object.
Similar to the issue in Bug 1 above.

Solution: Only select elements that are on_screen.