Improvement to object pathway
To improve the game, I made sure that any viable paths did not
contain obstacles. Previously this meant that new objects would
just fall to the first available top tile in that column. Now
new objects fall to the first top cell in the row that is not
an obstacle.

Amended level ten obstacles