Add wooden obstacles
The creation of wood obstacles are similar to that of Ice obstacles,
except that the player can not see the object underneath the obstacle.

Wood obstacles have to create a match twice before the urb is freed,
and the obstacle destroyed. So when the match happens the first time
the urb under the wood obstacle is removed from the match. The second
time the wood obstacle is destroyed and all urbs that make up that
match are freed.

Create level 12 that has wood obstacles.