Fix another problem with object movement
Specifically seen with levels 19 and 20. Some objects are moving over remaining objects in it's path to an empty cell.

This problem occurred when replacing new objects back into play. The new objects were going directly to the empty cells instead of finding a path that was free of objects or obstacles. This specific problem appears to be fixed.