Wooden obstacle bug
If there was a match with two wooden obstacles at
intersecting points e.g. T or L shape, the objects
underneath the wooden obstacle at the intersecting
point would bounce out of the obstacle immediately
although the counter is not yet zero and the
obstacle is damaged but not broken.

After investigation it was apparent that the match
was not combining these two matches into T or L shape,
but treating them as two separate matches. Also the
match was not checking for combined matches if the
object status was anything other than :NONE. Meaning
it was not analysing :WOOD or :CEMENT statuses.

These issues have been fixed