Determine viable shuffle objects
The process is not required if there are no
obstacles that are blocking movement on the
tilemap. But where there are obstacles that
are blocking entire rows or columns. The
movement of the objects must be restricted
to within the section they are blocked in.

Steps to completing this are:
Identify if the tilemap has blocked row and
how many blocked rows

Identify if the tilemap has blocked columns
and how many

Identify the sections that objects can move
within

Separate objects that are in each blocked
section

Remove any obstacles (invisible, ice/wood
etc)

Shuffle the objects within their blocked
boundary