Fix bug and swap shape definition
On 5 or more in a row the intersected element was showing nil
instead of a location of where the special treat will appear.
This has now been fixed.

Cookie now is generated on 5 or more in a row. Gobstopper is
generated when the match is a T or L shape.